Moved all code from ListTrait to BaseCsvListFieldHandler
ListTrait code is all about handling list values from the CSV files.
This should be internal to the list handlers and not visible from the
outside.

Now that we especially support multiple lists based on CSV (list and
sub-list) and list not even based on CSV (dblist), this functionality
is way to specific to be spilled outside BaseCsvListFieldHandler.
